+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog ##
 #### dev ####
+* Include Kotlin in ModTheSpire
 
 #### v3.17.0 ####
 * No longer require patch classes and methods to be public